// src/app/models/MercanciaEncabezado.ts
export class MercanciaEncabezado{
     idMercancia: number;
     codigoMercancia: number;
     nombreProducto: string;
     cantidad: number;
     fechaIngreso: Date;
     idUsuarioModifica: number;
     fechaModificacion: Date;
     usuarioCrea: string;
     usuarioModifica: string;
     estado: string;
     idUsuarioCrea: number;
}